+++ node/env0-deploy-cli.js
@@ -1,6 +1,6 @@
 const commandLineArgs = require('command-line-args');
 const commandLineUsage = require('command-line-usage');
-const runDeployment = require('./env0-deploy-front-end');
+const runDeployment = require('./env0-deploy-flow');
 
 const optionDefinitions = [
   { name: 'apiKey', alias: 'k', type: String, description: 'Your env0 API Key' },
@@ -10,19 +10,25 @@ const optionDefinitions = [
   { name: 'projectId', alias: 'p', type: String, description: 'Your env0 Project id' },
   { name: 'blueprintId', alias: 'b', type: String, description: 'The Blueprint id you would like to deploy with' },
   { name: 'environmentName', alias: 'e', type: String, description: 'The environment name you would like to create, if it exists it will deploy to that environment' },
+  { name: 'environmentVariables', alias: 'ce', type: String, multiple: true, description: 'The environment variables to set on the deployed environment - works only on deploy and can be multiple, the format is \"environmentVariableName1=value\"' },
   { name: 'revision', alias: 'r', type: String, defaultValue: 'master', description: 'You git revision, can be a branch tag or a commit hash. Default value \"master\" ' },
   { name: 'help', alias: 'h', type: Boolean, defaultValue: false, description: 'Get help' }
 ];
 
-const header = `
-                                         .oooo.   
-                                        d8P'\`Y8b  
- .ooooo.  ooo. .oo.   oooo    ooo      888    888 
-d88' \`88b \`888P"Y88b   \`88.  .8'       888    888 
-888ooo888  888   888    \`88..8'        888    888 
-888    .o  888   888     \`888'         \`88b  d88' 
-\`Y8bod8P' o888o o888o     \`8'           \`Y8bd8P'  
-`;
+const header = `                                                                                                          
+     ///(((///*****    ,,,,*                                                               ,%%%%%%%%%%%,     
+     ///(((///*****    ,,,,*                                                              #%%*       /%%/    
+                                                                                          %%%         #%%    
+                                     .#%%%%%%(       .%%* (%%%%%#.     #%%         #%%    %%%         #%%    
+        .                          #%%%*..../%%%,    .%%%%(,..,(%%%,    %%(       .%%*    %%%         #%%    
+     ///((    //***    **,,,       %%/        %%%    .%%/        %%%    *%%       %%%     %%%  ****   #%%    
+     ///((    //***    **,,,      .%%,        #%%    .%%,        %%%     #%%     /%%      %%%  ****   #%%    
+     ///((                        .%%%%%%%%%%%%%%    .%%,        %%%      %%,    %%.      %%%         #%%    
+     ///((                        .%%*               .%%,        %%%      *%%   #%%       %%%         #%%    
+     ///((    **********,,,*      .%%*               .%%,        %%%       %%% /%%        %%%         %%%    
+     ///((    **********,,,*       %%%               .%%,        %%%        %%.%%*        (%%#       %%%,    
+     ///((    **********,,,*        .%%%%%%%%%%%%    .%%,        %%%        /%%%%           #%%%%%%%%%#      
+                                                                                                             `;
 
 const sections = [{
     content: header,
